hmh-backend/requirements.txt
# HMH Backend — Python dependencies
# Python 3.12+  |  PostgreSQL 15+

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0         # async driver (for future async upgrade path)

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and settings
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20  # form data / file uploads

# Utilities
python-dotenv==1.0.1
httpx==0.28.1             # test client and outbound HTTP

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
